requirements.txt
asgiref==3.4.1
Django==3.2.23
django-cors-headers==3.10.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.4.0
mysqlclient==1.4.6
Pillow==8.4.0
PyJWT==2.4.0
pytz==2023.3.post1
sqlparse==0.4.4
typing-extensions==4.1.1
